+++ src/entities/question-button/index.ts
@@ -1,0 +1,1 @@
+export * from "./ui/question-button.ui";
