--- src/entities/question_button/index.ts
@@ -1,1 +1,0 @@
-export * from "./ui/question_button";
