+++ src/pages/root/root.page.tsx
@@ -1,11 +1,11 @@
 import styled from "styled-components";
 
-import { Top } from "~/widgets/Top";
 import { Header } from "~/widgets/header";
 import { LanguageRank } from "~/widgets/language-rank";
 import { QuestionGuide } from "~/widgets/question-guide";
 import { QuestionList } from "~/widgets/question-list";
 import { ReviewerRank } from "~/widgets/reviewer-rank";
+import { Top } from "~/widgets/top";
 
 import { Banner } from "~/shared/banner";
