+++ src/widgets/Top/ui/top.ui.tsx
@@ -1,11 +1,11 @@
 import styled from "styled-components";
 
 import { Filter } from "~/entities/filter";
-import { QuestionButton } from "~/entities/question_button";
+import { QuestionButton } from "~/entities/question-button";
 
 const Wrapper = styled.div`
   width: 100%;
-  max-width: 1440px;
+  height: 72px;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -13,6 +13,9 @@ const Wrapper = styled.div`
   gap: 100px;
   box-shadow: 0px 5px 10px rgba(117, 117, 121, 0.4);
   margin: auto;
+  & > * {
+    max-width: 1440px;
+  }
 `;
 
 export const Top = ({ isWritePage = false }: { isWritePage?: boolean }) => {
